Resolve casting related warnings

+++ include/raytracer/util/Surface.h
@@ -90,7 +90,7 @@ class Surface {
 		/**
 		 * Ray parameter.
 		 */
-		float m_T;
+		double m_T;
 
 		/**
 		 * TODO
